rename Metrics namespace to Provider
